+++ src/app/(auth)/sign-in/page.tsx
@@ -1,35 +1,10 @@
-'use client'
-
-import { Button } from '@/components/elements/button'
-import { ErrorBanner } from '@/components/elements/error-banner'
-import { InputField } from '@/components/elements/fields/input-field'
+import { Suspense } from 'react'
 import { GitHubSignInButton } from '@/components/features/auth/github-sign-in-button'
-import type { ActionState } from '@/types/form'
-import Link from 'next/link'
-import { useRouter } from 'next/navigation'
-import { useActionState } from 'react'
-import {
-  signInAction,
-  signInGithubAction,
-} from '../../../lib/server-actions/auth/sign-in-actions'
-import { toast } from 'sonner'
+import { signInGithubAction } from '../../../lib/server-actions/auth/sign-in-actions'
+import { GitHubSignInError } from '@/components/features/auth/github-sign-in-error'
+import { SignInForm } from '@/components/features/auth/sign-in-form'
 
 export default function SignInPage() {
-  const router = useRouter()
-
-  // 通常認証の状態管理（バリデーションエラー表示のため）
-  const [state, action, isPending] = useActionState(
-    async (state: ActionState | undefined, formData: FormData) => {
-      const result = await signInAction(formData)
-      if (result.success) {
-        toast.success('サインインしました')
-        router.push('/')
-      }
-      return result
-    },
-    undefined,
-  )
-
   return (
     <div className="flex justify-center items-center min-h-screen">
       <div className="relative w-full max-w-md rounded-lg bg-white p-8 border">
@@ -39,13 +14,17 @@ export default function SignInPage() {
           </h2>
         </div>
 
-        {state?.error && <ErrorBanner message={state.error.message} />}
-        
+        {/* GitHub認証セクション */}
+        <div className="mt-6 space-y-3">
+          <form action={signInGithubAction}>
+            <GitHubSignInButton />
+          </form>
 
-        {/* GitHub認証ボタン */}
-        <form action={signInGithubAction} className="mt-6">
-          <GitHubSignInButton />
-        </form>
+          {/* エラー表示（Suspenseで囲む） */}
+          <Suspense fallback={null}>
+            <GitHubSignInError />
+          </Suspense>
+        </div>
 
         {/* 区切り線 */}
         <div className="mt-6 mb-6">
@@ -59,42 +38,8 @@ export default function SignInPage() {
           </div>
         </div>
 
-        <form className="space-y-6" action={action}>
-          <div className="space-y-4">
-            <InputField
-              id="username"
-              label="メールアドレス"
-              type="email"
-              placeholder="メールアドレスを入力"
-              defaultValue={state?.values?.username}
-              errors={state?.error?.fields?.username}
-            />
-            <InputField
-              id="password"
-              label="パスワード"
-              type="password"
-              placeholder="パスワードを入力"
-              defaultValue={state?.values?.password}
-              errors={state?.error?.fields?.password}
-            />
-          </div>
-
-          <div>
-            <Button type="submit" disabled={isPending}>
-              サインイン
-            </Button>
-          </div>
-
-          <div className="mt-4 text-center text-sm text-gray-600">
-            アカウントをお持ちでないですか？{' '}
-            <Link
-              href="/sign-up"
-              className="font-medium text-indigo-600 hover:text-indigo-500"
-            >
-              新規登録
-            </Link>
-          </div>
-        </form>
+        {/* 通常のサインインフォーム */}
+        <SignInForm />
       </div>
     </div>
   )
